Moved actual action appliers into something called action reducers for better testability
IOW: Action code is part of the business logic and should be as reusable as possible.
It used to be tied to Redux. Now it is just a list of functions that alter the state,
which gives better reusability (if you want to use it without Redux) and better
testability (no need to init a store to use the code, you can just pass a state
object and test that the code interacts with the state as expected)